[HUDI-5681] Fixing Kryo being instantiated w/ invalid `SparkConf` (#7821)
This is addressing misconfiguration of the Kryo object used specifically to serialize Spark's internal structures (like `Expression`s): previously we're using default `SparkConf` instance to configure it, while instead we should have used the one provided by `SparkEnv`
